+++ packages/react-native-sdk/src/hooks/push/useIosVoipPushEventsSetupEffect.ts
@@ -2,13 +2,14 @@ import { type MutableRefObject, useEffect, useRef, useState } from 'react';
 import { getVoipPushNotificationLib } from '../../utils/push/libs';
 
 import { Platform } from 'react-native';
-import { onVoipNotificationReceived, StreamVideoRN } from '../../utils';
+import { StreamVideoRN } from '../../utils';
 import {
   useConnectedUser,
   useStreamVideoClient,
 } from '@stream-io/video-react-bindings';
 import { setPushLogoutCallback } from '../../utils/internal/pushLogoutCallback';
 import { StreamVideoClient, getLogger } from '@stream-io/video-client';
+import { onVoipNotificationReceived } from '../../utils/push/internal/ios';
 
 const logger = getLogger(['useIosVoipPushEventsSetupEffect']);
 
@@ -86,7 +87,7 @@ export const useIosVoipPushEventsSetupEffect = () => {
   useEffect(() => {
     const pushConfig = StreamVideoRN.getConfig().push;
     const pushProviderName = pushConfig?.ios.pushProviderName;
-    if (Platform.OS !== 'ios' || !pushConfig || !client || !pushProviderName) {
+    if (Platform.OS !== 'ios' || !client || !pushProviderName) {
       return;
     }
 
@@ -149,7 +150,7 @@ export const useIosVoipPushEventsSetupEffect = () => {
         if (name === 'RNVoipPushRemoteNotificationsRegisteredEvent') {
           onTokenReceived(data);
         } else if (name === 'RNVoipPushRemoteNotificationReceivedEvent') {
-          onVoipNotificationReceived(data);
+          onVoipNotificationReceived(data, pushConfig);
         }
       }
     });
